fix #7: user search parameters are not kept over sessions
If users do a search in the backend module, the search parameter gets
saved in the backend user configuration. This is needed as the users
want to keep their view on the module especially if plenty of categories
and contacts are present.

This feature got lost by code refactoring in a previous commit.