Fix issue were login message would appear for logged in users
The problem was that users that are unknown by SLUB_Events would get a
login message even though they are logged in.

The check for {subscribers} only works if the user previously was
subscribed to any event. Only then he/she is in the database table
`tx_slubevents_domain_model_subscriber` and the first if-condition is
not null.

Therefore we have to do a two-step check to display the message to avoid
showing the hint to logged in users. They get the info that they are not
subscribed to any event yet instead.