Move required attribute to additionalAttributes.
The TextareaViewhelper knows the required attribute as of TYPO3 9.2:

[Feature: #82704 - Add readonly and required attributes to
TextareaViewHelper](https://docs.typo3.org/c/typo3/cms-core/10.2/en-us/Changelog/9.2/Feature-82704-AddReadonlyAndRequiredAttributesToTextareaViewHelper.html)

We still need the compatibilty with TYPO3 8.7 LTS and do not which to
provide our own TextareaViewHelper as before. That's why the required
attribute moved to additionalAttributes.

This fixes #39.